+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/state/content/DownloadState.kt
@@ -16,6 +16,8 @@ import kotlin.random.Random
  * @property fileName A canonical filename for this download.
  * @property contentType Content type (MIME type) to indicate the media type of the download.
  * @property contentLength The file size reported by the server.
+ * @property currentBytesCopied The number of current bytes copied.
+ * @property status The current status of the download.
  * @property userAgent The user agent to be used for the download.
  * @property destinationDirectory The matching destination directory for this type of download.
  * @property filePath The file path the file was saved at.
@@ -32,6 +34,8 @@ data class DownloadState(
     val fileName: String? = null,
     val contentType: String? = null,
     val contentLength: Long? = null,
+    val currentBytesCopied: Long = 0,
+    val status: Status = Status.INITIATED,
     val userAgent: String? = null,
     val destinationDirectory: String = Environment.DIRECTORY_DOWNLOADS,
     val referrerUrl: String? = null,
@@ -41,4 +45,35 @@ data class DownloadState(
 ) : Parcelable {
     val filePath: String get() =
         Environment.getExternalStoragePublicDirectory(destinationDirectory).path + "/" + fileName
+
+    /**
+     * Status that represents every state that a download can be in.
+     */
+    enum class Status {
+        /**
+         * Indicates that the download is in the first state after creation but not yet [DOWNLOADING].
+         */
+        INITIATED,
+        /**
+         * Indicates that an [INITIATED] download is now actively being downloaded.
+         */
+        DOWNLOADING,
+        /**
+         * Indicates that the download that has been [DOWNLOADING] has been paused.
+         */
+        PAUSED,
+        /**
+         * Indicates that the download that has been [DOWNLOADING] has been cancelled.
+         */
+        CANCELLED,
+        /**
+         * Indicates that the download that has been [DOWNLOADING] has moved to failed because
+         * something unexpected has happened.
+         */
+        FAILED,
+        /**
+         * Indicates that the [DOWNLOADING] download has been completed.
+         */
+        COMPLETED
+    }
 }