+++ mobile/android/android-components/components/feature/downloads/src/test/java/mozilla/components/feature/downloads/manager/AndroidDownloadManagerTest.kt
@@ -24,7 +24,6 @@ import org.mockito.ArgumentMatchers.anyString
 import org.mockito.Mockito.verify
 import org.mockito.Mockito.verifyZeroInteractions
 import mozilla.components.feature.downloads.AbstractFetchDownloadService.Companion.EXTRA_DOWNLOAD_STATUS
-import mozilla.components.feature.downloads.AbstractFetchDownloadService.DownloadJobStatus
 import mozilla.components.support.test.libstate.ext.waitUntilIdle
 import org.junit.Assert.assertEquals
 
@@ -40,7 +39,7 @@ class AndroidDownloadManagerTest {
         download = DownloadState(
             "http://ipv4.download.thinkbroadband.com/5MB.zip",
             "", "application/zip", 5242880,
-            "Mozilla/5.0 (Linux; Android 7.1.1) AppleWebKit/537.36 (KHTML, like Gecko) Version/4.0 Focus/8.0 Chrome/69.0.3497.100 Mobile Safari/537.36"
+            userAgent = "Mozilla/5.0 (Linux; Android 7.1.1) AppleWebKit/537.36 (KHTML, like Gecko) Version/4.0 Focus/8.0 Chrome/69.0.3497.100 Mobile Safari/537.36"
         )
         store = BrowserStore()
         downloadManager = AndroidDownloadManager(testContext, store)
@@ -98,7 +97,7 @@ class AndroidDownloadManagerTest {
     @Test
     fun `sendBroadcast with valid downloadID must call onDownloadStopped after download`() {
         var downloadCompleted = false
-        var downloadStatus: DownloadJobStatus? = null
+        var downloadStatus: DownloadState.Status? = null
         val downloadWithFileName = download.copy(fileName = "5MB.zip")
 
         grantPermissions()
@@ -117,12 +116,12 @@ class AndroidDownloadManagerTest {
         notifyDownloadCompleted(id)
 
         assertTrue(downloadCompleted)
-        assertEquals(DownloadJobStatus.COMPLETED, downloadStatus)
+        assertEquals(DownloadState.Status.COMPLETED, downloadStatus)
     }
 
     @Test
     fun `sendBroadcast with completed download removes queued download from store`() {
-        var downloadStatus: DownloadJobStatus? = null
+        var downloadStatus: DownloadState.Status? = null
         val downloadWithFileName = download.copy(fileName = "5MB.zip")
         grantPermissions()
 
@@ -139,7 +138,7 @@ class AndroidDownloadManagerTest {
 
         notifyDownloadCompleted(id)
         store.waitUntilIdle()
-        assertEquals(DownloadJobStatus.COMPLETED, downloadStatus)
+        assertEquals(DownloadState.Status.COMPLETED, downloadStatus)
         assertTrue(store.state.queuedDownloads.isEmpty())
     }
 
@@ -165,14 +164,14 @@ class AndroidDownloadManagerTest {
     private fun notifyDownloadFailed(id: Long) {
         val intent = Intent(ACTION_DOWNLOAD_COMPLETE)
         intent.putExtra(android.app.DownloadManager.EXTRA_DOWNLOAD_ID, id)
-        intent.putExtra(EXTRA_DOWNLOAD_STATUS, DownloadJobStatus.FAILED)
+        intent.putExtra(EXTRA_DOWNLOAD_STATUS, DownloadState.Status.FAILED)
         testContext.sendBroadcast(intent)
     }
 
     private fun notifyDownloadCompleted(id: Long) {
         val intent = Intent(ACTION_DOWNLOAD_COMPLETE)
         intent.putExtra(android.app.DownloadManager.EXTRA_DOWNLOAD_ID, id)
-        intent.putExtra(EXTRA_DOWNLOAD_STATUS, DownloadJobStatus.COMPLETED)
+        intent.putExtra(EXTRA_DOWNLOAD_STATUS, DownloadState.Status.COMPLETED)
         testContext.sendBroadcast(intent)
     }